fixed the bug
The naive heap didn't actually remove entries when they were
extracted, so running Dijkstra's algorithm with it never finished.

The behavior of the program stopping happened due to this infinite
loop.

A secondary bug was that, during testing, I accidentally had two
handlers, with the same observable behavior, registered at the same
time.

That caused doubled output when using a routine that did complete.

Both bugs are fixed now.