wrote Consolidate and its helpers
This is quite rough, likely should be refactored, hopefully is
correct but I'm not confident, and I haven't even actually compiled
it, since other things block that for the moment (for example, the
IPriorityQueue interface is not fully implemented).

I'll want to revisit the structure of this code and see if I can
have it make more sense (and perhaps decompose stuff into smaller
pieces). But I think I'll first move on to implementing "decrease
key" and the associated map data structure.

Once it's working, I can improve it. Or if it doesn't work and I
can't figure out why, my consternation may help point to a better
design.