did initial refactoring work
This builds and runs, but its behavior is hardly different.

Neither of the two goals of this refactoring effort have yet been
achieved. This is just groundwork.

There is also the question of whether this degree of reliance on
reflection, for something that *can* be achieved without it,
is a good idea.