implemented EdgeSelection
But I don't think ComputeShortestPaths should actually return an
EdgeSelection object, because the parents tree that Dijkstra's
algorithm naturally produces (which ComputeShortestPaths currently
returns) it useful in and of itself.

Perhaps there should also be a Tree or ParentsTree class, with a
ToEdgeSelection method, and ComputeShortestPaths can return a
Tree/ParentsTree object.