allow custom informal names
This will be important when the Fibonacci heap is implemented,
since forming it from its type name automatically would cause the
informal name to be "fibonacci heap" (uncapitalized) when it should
be "Fibonacci heap" (captalized).

This is tested/demonstrated by giving a custom informal name to
UnsortedArrayPriorityQueue, but that is not actually desirable.