use #load instead of reproducing the implementation
Since, at least for now, I'm no longer making special modifications
to the scratchwork version.

The big question on my mind is why it seems to work, in the actual
program, even though Decrease() is not yet implemented. To
investigate that, I'll want debugging code (specifically dumping)
that both the Dijkstra's algorithm implementation and the scratch
tests cause to be run.