healed various error-reporting papercuts
The following nuisances were fixed:

(1) Graphs of order 0 must have size 0 even if parallel edges are
allowed.

(2) An out-of-range integer should be reported as out of range,
not in a way that confusingly implies it's not an integer at all.

Intervals with out-of-range endpoints should be treated specially,
too, but I have not done that (yet).

(3) Weight errors, other than not specifying weights at all (i.e.,
a empty/whitespace weight specification), should be reported after
other errors, since the weight text box (and the related check box)
are lower down in the dialog box than most other stuff.

Also, some code that was hard to read was made clearer, and an
integer overflow bug that couldn't occur *just* because of the
specific way this application happens to use its data structures
was fixed.