implemented GraphGenerator
Though there will probably need to be some further refinements, and
it is only very minimally tested.

Note that GraphGeneratorDialog is not complete, because it does not
yet facilitate subscription to GraphGenerator results. It's only
the backend (or middle-end and back-end, depending how you slice
it), i.e., GraphGenerator itself, that might be done.