implemented "system menu" functionality
But there is a regression, where pressing F7 when the caret is
enabled doesn't always disable it.

I'm not sure if that's new as of this commit or is older. It might
have come in at d76b8f5.

Update: It looks like this bug has always been present, or at least
for a long time. Clicking another textbox (or checkbox) and then
back in the status textbox shows the caret, even if the caret isn't
supposed to be shown.