fixed bug where sizes were checked against max order instead of min
I apparently missed this in the testing I did a while ago, because
the inputs that produce it don't *always* produce it (since a
higher order may be randomly selected). If I'd tested with smaller
ranges of multiple values, I would've found it before.

An example that produced the bug often enough that it can
effectively be used to reproduce it in affected versions (just
click Generate several times until the unhandled exception is
thrown) is:

    Order: 3-5
    Size: 7-10
    Allow loops: no
    Allow parallel edges: no

(Anything else for the other settings.)

That now rightly produces the error status message:

    3 vertices allow only 6 edges