respect display scaling
This turns on automatic size adjustment based on the display
settings encountered at runtime. Without this, the graph generator
dialog window is completely broken and inaccessible on numerous
systems, and only displays correctly with a scaling of 100%.

The technique used is analogous to the approach taken in code
generated by the Windows Forms designer (though I have not copied
code from that, nor have I organized this code in a similar way).

The changes to accommodate > 100% scaling are not finished. The
one bug I've noticed is that "unique edge weights" renders as
"unique edge" when the scale is 150%. I can stretch the UI out to
prevent this, but I expect problems like this to occur at even
higher scalings. Stretching the interface so much to prevent it at
any likely scaling would greatly harm the interface at low
scalings.

A scaling of 250% is common on small Ultra HD displays, such as
laptops with 2160p screens.

The other, more trivial way that this change is incomplete is that
various comments showing things that didn't need to be (and usually
shouldn't be) done remain, and some debugging output is being
printed.