widened status box
It was accidentally smaller than it had to be.

This makes it a bit taller too, to match the height of the
textboxes. But this doesn't seem to make a difference. In
particular, descenders (like the curve at the bottom of the "g" in
"weights") are still, slightly but apparently by the same amount,
cut off at high scalings.

At this point, the UI looks reasonable at scalings up to 175%.

Bigger scalings are possible and even common, on Ultra HD displays.
I'll have to go to another machine to test that though.