completed Fibonacci heap implementation
This commit adds implementation code for Decrease and its helper
functions, Cut and CascadingCut.

Two important notes:

(1) This does not actually have all the Fibonacci heap operations,
just the ones needed for Dijkstra's algorithm. In particular,
Fibonacci heaps are efficiently mergeable, but no merge operation
is provided here.

(2) This is only very lightly tested so far, and it may well have
bugs. I've tested it with the two test inputs included on this
branch, and as expected it failed before those functions were
implemented and passed once they were.