fixed the "<limit of graph>" bug on Clear
This is not necessarily the best fix, but it is probably the most
minimal of all the reasonable ways to fix it.

Ultimately, I should re-evaluate the way the controller and output
are displayed, and likely use a DumpContainer or something (though
I believe that would prevent me from showing nice headings as I'm
doing now).

It would also, or instead, be good to use a better control than a
LINQPad.Controls.TextArea for the textarea. If possible, I'd like
to use https://github.com/Microsoft/monaco-editor, though I think
I'd have to implement a wrapper for it that inherits from
LINQPad.Controls.Control to marshal data and and out of JavaScript.

Besides (presumably) much better performance for large amounts of
text, this would also have the benefit of supporting IntelliSense-
like red and green squiggles on input that is syntactically wrong
or contains vertices not in the range [0, order) -- though of
course I'd have to implement the logic for that.

But even assuming I do that, I'll likely want this program to
continue to work without that dependency, and possibly even to
allow the user to decide whether or not to use it.

I'd sacrifice having both available if having just Monaco would
be far less complex, though.