failed attempt to call base WndProc only when needed
The problem here is that lots of other stuff is signalled via
WM_SYSCOMMAND: other system menu items, but also opening the system
menu in the first place, moving the window by dragging the title
bar, etc.

So the old way was correct, it just should have a comment
explaining why.