an async experiment (not keeping this)
If the caller (who creates the Controller object, in this case the
caller is Main) is going to decide how computations are performed
by subscribing handler to the SingleRun and RunsCompleted events,
then the caller, and not Controller, should be responsible for
deciding what runs asynchronously and how it runs (e.g., whether or
not to start multiple tasks before awaiting them so as to allow
them to run concurrently).

So this design, which keeps--and further solidifies--that specific
event-based design, yet has the Controller wrap calls to those
events' handlers in tasks and runs them asynchronously-- should not
be used.

I wrote "FIXME" comments that assumed that this design--this
attempt to restrict the purview of the Controller--would be kept.
But I'm not sure I should keep it. Intuitively, it seems like the
Controller should be responsible for more, and that seems easier.

Right now, the Controller keeps track of how priority queues of
specific types are to be created. This is an unusual thing to do if
the caller is supposed to be responsible for deciding what
computations to perform. Of course, I can refactor/redesign so the
Controller doesn't hold references to priority queue suppliers.
But this does show a long-standing lack of committment on my part
in this code base to having the Controller be limited in this way.

Furthermore, it makes sense for the Controller in dijkstra.linq to
serve a purpose analogous to that of GraphGeneratorDialog in
generator.linq. If its purpose is broadened to do that, then it
will decide what computations to perform as well as what results
are desired, but will perhaps not be responsible for deciding what
to do with the results themselves.

Or it could be broader and take care of that too.

There's some risk here that the Controller will become a "God
object" but the important parts of the program--the implementation
of Dijkstra's algorithm, the IPriorityQueue interface, and the
types that implement it--are all separate from the Controller and
have no dependency on it. So I think it may turn out okay.

The other big question, regarding this experiment, is: did it work?
Does asynchronicity here improve responsiveness? The answer, as far
as I can tell, is, "sort of." Latency from storing enough edges in
the textarea to do big tests (millions of edges), and possibly
effects of paging during big memory allocations, make it hard to
tell. Also, right now, generating the Graph object from the edges
is still synchronous.

The Graph Generator tool window still stops responding during
substantial parts of processing. Testing in lag-scratch.linq
does not suggest a reason that would happen in the current code
(the code in this commit that has async stapled in temporarily),
and I expect a properly implemented version not really to be better
in terms of lag. I wonder if dumping is doing some work on the UI
thread in spite of my best efforts. It might be useful to look into
attempting to use a separate UI thread for tool windows (of which
currently there is only one, GraphGeneratorDialog).

Finally, in a test with a dense graph of 4000 vertices and about
16 million edges, about 6 GiB of RAM were taken up by LINQPad (not
the query, but LINQPad itself) when populating the textarea, which
dropped down to about 4.5 GiB afterwards (I'm talking about memory
obtained from, and released back to, the system). As for the query,
the graph generator took up about 2 GiB. This is a lot of overhead.
It may be good to profile to see, but I'm pretty sure this is due
to the space overhead of using a hash table (Dictionary<,>) to
implement the selection-without-repetitions algorithm. That can be
made to use less memory in two ways: aggressively remove each key
when it is guaranteed not to be accessed again, and for some
threashold of range and sample count (i.e., when the sample count
is near enough to the full range), use an array (really a List<,>)
instead of a hash table (Dictionary<,>).

The main point of this program is to demonstrate how Dijkstra's
algorithm can use various priority queues and can be well
understood as an algorithm that leverages an abstract priority
queue, and to draw pretty pictures via Graphviz (possibly also via
other means, like LaTeX and MSAGL/automatic-graph-layout). Still,
it would be nice if it worked well on big graphs too. So far, it
exhibits substantial UI unpleasantness, as well as greater memory
usage than should be necessary, at sizes far less than those that
make performance comparisons between priority queue data structures
interesting. If would be nice to fix this. In addition to more and
better asynchronicity and possibly an constant-factor space
improvement on the edge generation algorithm, it may (as I've
mentioned before) be good to use a different control than a
LINQPad.Controls.TextArea (which wraps a JavaScript textarea in
the WFP WebBrowser control, which itself wraps the WebBrowser
ActiveX object that is effectively Internet Explorer). Assuming it
works in the WebBrowser control and that I can manage to wrap it
(being in JavaScript) in C# (inheriting from LINQPad.Control), I'd
like this to be Monaco. I'd still like to let the user choose
between the two, if that is feasible. Alternatively, it may be that
this program, at least in its current design, has ougrown LINQPad
and should be implemented as a Blazor app (likely a progressive
webapp) or Avalonia app, or something. It would be nice for it to
work on non-Windows OSes too, which both of those would allow.