fire run_Click on a worker thread (not keeping this)
This change, by itself, is sufficient to remedy the current lag.

However, this is still not a very good approach. Moving control of
asynchronous execution all into the Controller with the careful use
of await (or all outside the Controller, also with await, which is
the fix I had originally thought best) should also allow steps that
don't depend on one another to be run concurrently, and for the
use to control whether or not (and perhaps even specifically when)
this is done.

Note that even generating the graph from the edge list in the
textarea will have to be done asynchronously--at least when the
list is long--in order to avoid lag.

If that is ultimately untenable or needs to be postponed, then the
forthcoming commit that reverts this can itself be reverted.
Another alternative in that case would be to run tool windows (of
which there is currently just one, the graph generator dialog) on
their own separate UI thread (by creating a thread, *not* using
a thread-pool thread, setting its apartment state to STA, and
using Windows.Forms.Application.Run to start a separate event loop
on it).