idea for fine-grained async
This works, but it doesn't give totally smooth operation, even in
terms of lag in a tool window (which, unlike the LINQPad UI,
*should* have absolutely no perceptible lag during any operation).

See the comments in run_Click, especially the first comment.

Summary of that comment: I'll probably go back to having the _run
button invoke its Click event handler asynchronously on a
threadpool worker thread, by setting its IsMultithreaded property
to true. But I'd like to better understand why that is effective.

The cases where I would still use async and await are:

(1) Partially for the dot process. Interacting with the streams
would still be asynchronous, since that seems to be the cleanest
way. But I could call WaitForExit to block until the process
terminates. In general it is bad to block a worker thread for I/O,
but at most one worker thread would ever be blocked for this reason
in this program. The only problem there is that if I ever make the
separate dot processes for separate groups (when not all results
agree, so there is more than one group) run at the same time, then
I'd need to change the way I do it or I'd be blocking potentially
too many worker threads.

(2) Running Dijkstra's algorithm with the different priority queues
concurrently (in "parallel") rather than in series as I now do. I'd
like to provide a runtime option for the user to decide if this is
done.

As for the other code changes, I'm interested in possibly
continuing with them (not leaving them in this shabby state, which
overuses tuples even by my standards) for two reasons:

(1) The old design does not make it easy to add concurrency between
opeations that don't depend on each other (in partcular, running
Dijkstra's algorithm concurrently with the different priority
queues).

(2) In the old design, results is a local variable, but due to the
particular way it is captured and used, it is effectively almost a
global variable... and there is an annoying bug, where it is not
cleared if a run throws an exception, so a subsequent run wrongly
reports inconsistent results due to the left-over data. This can be
fixed in other ways (such as by adding an event that triggers just
before Dijkstra's algorithm is run, to mirror the event that
triggers afterwards, and clearing results in that) but if changes
are to be made to fix it, they may as well make the design less
brittle.