made the DOT code box use a monospaced font (since it holds code)
I'm not sure this change is good. I had to make it wider too, to
avoid unsightly wrapping in common (actually, currently all) cases.
Now it's looks about as big as the edges box, which doesn't seem
right. Also, the edges box only accepts text in a very specific
format, except that extra whitespace and #-comments are supported,
so it's pretty code-like too. Should it be monospace a well?